Handwired/Dactyl keyboard (#2058)
* Copy the ergodox_ez code to handwired/dactyl

Differences from the Ergodox:

* Use QMK_SUBPROJECT_H instead of QMK_KEYBOARD_H, since it's under
    handwired
* Omitted several keymaps. They'll eventually be broken (since the
    Dactyl has fewer keys), and I don't want to try to fix them.
* Omitted the keymap images for the default layout, since they depict
    a different keyboard.
* Everything that said Ergodox now says Dactyl, naturally.

* [whitespace] Delete trailing whitespace

My editor does this automatically so it's just gonna keep cropping up...

* Cut the dactyl down to the right number of columns

(Remember, throughout matrix.c, everything called "row" is really a
column, and vice-versa).

* Remove LED-related code

* Tighten up the Dactyl's build options

* Whitespace cleanup in twimaster.c

* Hardtabs -> spaces
* No more trailing whitespace

* Typo fix

* Correct the CPU frequency units

The Teensy's CPU definitely doesn't run at 16 petahertz...

* Restore access to ONEHAND_ENABLE

I turned it off in 26d47cb42622d990a7c3335e7fcc151aa3edfbf0 while
desperately debugging; I just wanted to ensure it wasn't causing the
problem I was seeing. It was not, in fact, causing the problem, so it's
back.

Also fixed the swap matrix in dactyl.c, since it still referred to
columns that exist in the Ergodox but not the Dactyl.

* Clearer phrasing about TWI's effect on scan rate

* Fix up the Dactyl's firmware-loading instructions

Sadly, the Dactyl has no hole for the onboard reset button.

* Dvorak keymap for the Dactyl

* The Erincalling Layout

* Erincalling layout: Add a := key

I've been working in Go, which uses := a lot, and it's awkward to type
in this layout.

* Dactyl README: link to the dactyl-keyboard repo

* Add a missing copyright line

I don't know how much this matters? Honestly, it's enough for me that my
name is on the git commit. But hey, let's be consistent until there's a
specific reason not to be, right?

* Dactyl: remove commented-out code

I hate it I hate it I hate it

There's not even any information about what it was trying to do!!!! >:(

* Add a note about the row/column ridiculousness

* [whitespace] realign some constants

* Don't claim B4 is tied to VCC

It doesn't matter at all? I honestly don't know what the reason ever
was. It looks like it dates back to the original ErgoDox and I've never
seen one sentence about the purpose.

I've been skipping that wire for some time, and I promise it works fine.

* Dactyl keymaps: Send RALT for right-hand alt key

Not terribly important but I just like things tidy OK

* typo fix

* Refer to "dactyl.h" explicitly

QMK_SUBPROJECT_H has been working locally, but fails in CI. Strange!

* Dactyl: Don't use QMK_SUBPROJECT_H at all

It's still breaking in CI, even though it was a never a problem locally.